+++ backend/src/main/java/com/ladysparks/ttaenggrang/global/docs/EtfApiSpecification.java
@@ -20,23 +20,23 @@ public interface EtfApiSpecification {
 
     @Operation(summary = "ETF 상세 조회", description = "💡 ETF ID로 주식을 조회합니다.")
     @GetMapping
-    public ResponseEntity<EtfDTO> getEtf(@PathVariable("etfId") int etfId);
+    public ResponseEntity<EtfDTO> getEtf(@PathVariable("etfId") Long etfId);
 
     @Operation(summary = "ETF 매수", description = "💡 ETF ID와 수량으로 주식을 매수합니다.")
     @PostMapping("/{etfId}/buy")
-    public ResponseEntity<ApiResponse<EtfTransactionDTO>> buyEtf(@PathVariable("etfId") int etfId,
+    public ResponseEntity<ApiResponse<EtfTransactionDTO>> buyEtf(@PathVariable("etfId") Long etfId,
                                                                    @RequestParam("share_count") int shareCount,
                                                                    @RequestParam("studentId") Long studentId);
 
     @Operation(summary = "ETF 매도", description = "💡 ETF ID와 수량으로 주식을 매도합니다.")
     @PostMapping("/{etfId}/sell")
-    public ResponseEntity<ApiResponse<EtfTransactionDTO>> sellEtf(@PathVariable("etfId") int etfId,
+    public ResponseEntity<ApiResponse<EtfTransactionDTO>> sellEtf(@PathVariable("etfId") Long etfId,
                                                                       @RequestParam("share_count") int shareCount,
                                                                       @RequestParam("studentId") Long studentId);
 
     @Operation(summary = "변동률", description = "💡 주식 변동률 조회")
     public ResponseEntity<ApiResponse<EtfDTO>> updateEtfPrice(
-            @PathVariable("etfId") int etfId);
+            @PathVariable("etfId") Long etfId);
 
 //    @Operation(summary = "ETF 등록", description = "💡 새로운 ETF를 등록합니다.")
 //    @PostMapping("/create")
